+++ src/js/utils/algorithms/CatmullRomPath.d.ts
@@ -7,10 +7,13 @@
  * This demo implementation was inspired by this Codepen by Blake Bowen
  * https://codepen.io/osublake/pen/BowJed
  *
- * @date 2020-04-15
- * @author Converted to a class by Ikaros Kappler
+ * @date     2020-04-15
+ * @author   Converted to a class by Ikaros Kappler
  * @modified 2020-08-19 Ported this class from vanilla JS to TypeScript.
- * @version 1.0.1
+ * @version  1.0.1
+ *
+ * @file CatmullRomPath
+ * @public
  **/
 import { CubicBezierCurve } from "../../CubicBezierCurve";
 import { Vertex } from "../../Vertex";